change buffer length to unsigned int

+++ synth-plugin/dsp/include/DelayLine.h
@@ -21,7 +21,7 @@ public:
     void setDelay(float delayInSeconds);
     void write(float x);
     float read();
-    void processBlock(float* in, float* out, unsigned long length);
+    void processBlock(float* in, float* out, unsigned int length);
     void clear();
 };
 };
